Put functions separately into their own folder. Removed stage group and stage suffix from ajcc validation script.
